feat: add preloader functionality
Add a new feature of preloader, including relevant styles and scripts. Modify the scripts.ejs file to render 'js/libs/anime.min.js' only if the 'preloader' theme is not enabled. Extend the styles.css file with new classes related to preloader functionality. Adjust the _config.yml to enable the preloader. Add a new preloader.ejs file under '_partials'. Update the '_partials/head.ejs' file to include the preloader if it is enabled in the theme.