feat: implement responsive design in preloader
This commit introduces changes to the preloader.ejs file. It includes a media query to adjust the font size for screen sizes less than 600px to enhance the responsiveness of the preloader. The commit also adds padding to the preloader for better positioning and visual appeal.

In addition, the text animation is improved. The text is now split into words, each word and its letters are wrapped in spans, creating a more dynamic and visually appealing animation. A space is added between words for better readability.