feat: Add submenu toggle functionality and styles in navbar
This commit introduces a new feature that allows submenu toggling in the navbar. The feature is implemented both in the navbarShrink.js file and the navbar.ejs layout file.

In the JavaScript file, a new method called 'toggleSubmenu()' is introduced that listens for click events on elements with the attribute 'navbar-data-toggle'. The method controls the visibility of the target by toggling the 'hidden' class and also handles the rotation of the chevron icon.

In the navbar.ejs layout file, the structure of the navbar items is updated to accommodate the new submenu feature. The 'href' attribute of the anchor tag is conditionally set based on whether the current item has a submenu or not. The 'onClick' attribute is also conditionally added to prevent default behavior when the item has a submenu.

The commit also includes updates to the navbar.styl and styles.css files to style the new submenu feature. The transformations for the '