style: set max width for social links
This commit updates the styling of elements in home-banner.ejs and styles.css.
In home-banner.ejs, the width of the div containing social links has been
adjusted to max 1340px. In styles.css, various style properties such as
margin-top, height and max-width have been reordered and updated. This commit
also adjusts media query for font-size and line-height in styles.css.